requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.0
Django==5.1.2
Faker==30.8.2
gunicorn==23.0.0
packaging==24.2
pillow==11.0.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.1
tornado==6.4.1
typing_extensions==4.12.2
